Fix unaccesabele setings window Cleening
