+++ src/Camera/GetFrame/main.py
@@ -1,0 +1,20 @@
+from src.Camera.GetFrame.GetFrameUSB import GetFrameUSB
+from src.Camera.GetFrame.GetFreamFromCameraProducent import GetFrameFromProducent
+
+
+class GetFrame(GetFrameFromProducent, GetFrameUSB):
+    device = None
+
+    def __init__(self):
+        self.establishConnection()
+
+        if self.isConnectionEstablished:
+            GetFrameFromProducent.__init__(self)
+        else:
+            GetFrameUSB.__init__(self)
+
+    def getFrame(self):
+        if self.isConnectionEstablished:
+            return self.getFrameProducent()
+        else:
+            return self.getFrameUSB()
