+++ src/Camera/Setings/CameraSetingsFromProducent.py
@@ -4,7 +4,4 @@ from src.Camera.FromProducent.Abstract import AbstractCameraFromProducent
 class CameraSettingsFromProducent(AbstractCameraFromProducent):
 
     def showProducentSettings(self):
-        self.tisgrabber.IC_ShowPropertyDialog(self.handle)
-
-    def setWhiteBalanceAuto(self):
-        self.tisgrabber.IC_SetWhiteBalanceAuto(self.handle, 1)
+        self.tisgrabber.IC_ShowPropertyDialog(self.master.camera.handle)
